+++ xarray/core/dataset.py
@@ -5,6 +5,7 @@ import functools
 from collections import Mapping, defaultdict
 from distutils.version import LooseVersion
 from numbers import Number
+import warnings
 
 import sys
 
@@ -26,7 +27,7 @@ from .coordinates import DatasetCoordinates, LevelCoordinatesSource, Indexes
 from .common import ImplementsDatasetReduce, BaseDataObject
 from .dtypes import is_datetime_like
 from .merge import (dataset_update_method, dataset_merge_method,
-                    merge_data_and_coords)
+                    merge_data_and_coords, merge_variables)
 from .utils import (Frozen, SortedKeysDict, maybe_wrap_array, hashable,
                     decode_numpy_dict_values, ensure_us_time_resolution)
 from .variable import (Variable, as_variable, IndexVariable,
@@ -1126,6 +1127,78 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject,
                                  for k, v in self.variables.items()])
         return self._replace_vars_and_dims(variables)
 
+    def _validate_indexers(self, indexers):
+        """ Here we make sure
+        + indexer has a valid keys
+        + indexer is in a valid data type
+        """
+        from .dataarray import DataArray
+
+        invalid = [k for k in indexers if k not in self.dims]
+        if invalid:
+            raise ValueError("dimensions %r do not exist" % invalid)
+
+        # all indexers should be int, slice, np.ndarrays, or Variable
+        indexers_list = []
+        for k, v in iteritems(indexers):
+            if isinstance(v, integer_types + (slice, Variable)):
+                pass
+            elif isinstance(v, DataArray):
+                v = v.variable
+            elif isinstance(v, tuple):
+                v = as_variable(v)
+            elif isinstance(v, Dataset):
+                raise TypeError('cannot use a Dataset as an indexer')
+            else:
+                v = np.asarray(v)
+            indexers_list.append((k, v))
+        return indexers_list
+
+    def _get_indexers_coordinates(self, indexers):
+        """  Extract coordinates from indexers.
+        Returns an OrderedDict mapping from coordinate name to the
+        coordinate variable.
+
+        Only coordinate with a name different from any of self.variables will
+        be attached.
+        """
+        from .dataarray import DataArray
+
+        coord_list = []
+        for k, v in indexers.items():
+            if isinstance(v, DataArray):
+                v_coords = v.coords
+                if v.dtype.kind == 'b':
+                    if v.ndim != 1:  # we only support 1-d boolean array
+                        raise ValueError(
+                            '{:d}d-boolean array is used for indexing along '
+                            'dimension {!r}, but only 1d boolean arrays are '
+                            'supported.'.format(v.ndim, k))
+                    # Make sure in case of boolean DataArray, its
+                    # coordinate also should be indexed.
+                    v_coords = v[v.values.nonzero()[0]].coords
+
+                coord_list.append({d: v_coords[d].variable for d in v.coords})
+
+        # we don't need to call align() explicitly, because merge_variables
+        # already checks for exact alignment between dimension coordinates
+        coords = merge_variables(coord_list)
+
+        for k in self.dims:
+            # make sure there are not conflict in dimension coordinates
+            if (k in coords and k in self._variables and
+                    not coords[k].equals(self._variables[k])):
+                raise IndexError(
+                    'dimension coordinate {!r} conflicts between '
+                    'indexed and indexing objects:\n{}\nvs.\n{}'
+                    .format(k, self._variables[k], coords[k]))
+
+        attached_coords = OrderedDict()
+        for k, v in coords.items():  # silently drop the conflicted variables.
+            if k not in self._variables:
+                attached_coords[k] = v
+        return attached_coords
+
     def isel(self, drop=False, **indexers):
         """Returns a new dataset with each array indexed along the specified
         dimension(s).
@@ -1142,40 +1215,45 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject,
         **indexers : {dim: indexer, ...}
             Keyword arguments with names matching dimensions and values given
             by integers, slice objects or arrays.
+            indexer can be a integer, slice, array-like or DataArray.
+            If DataArrays are passed as indexers, xarray-style indexing will be
+            carried out. See :ref:`indexing` for the details.
 
         Returns
         -------
         obj : Dataset
             A new Dataset with the same contents as this dataset, except each
-            array and dimension is indexed by the appropriate indexers. In
-            general, each array's data will be a view of the array's data
-            in this dataset, unless numpy fancy indexing was triggered by using
+            array and dimension is indexed by the appropriate indexers.
+            If indexer DataArrays have coordinates that do not conflict with
+            this object, then these coordinates will be attached.
+            In general, each array's data will be a view of the array's data
+            in this dataset, unless vectorized indexing was triggered by using
             an array indexer, in which case the data will be a copy.
 
         See Also
         --------
         Dataset.sel
-        Dataset.sel_points
-        Dataset.isel_points
         DataArray.isel
         """
-        invalid = [k for k in indexers if k not in self.dims]
-        if invalid:
-            raise ValueError("dimensions %r do not exist" % invalid)
-
-        # all indexers should be int, slice or np.ndarrays
-        indexers = [(k, (np.asarray(v)
-                         if not isinstance(v, integer_types + (slice,))
-                         else v))
-                    for k, v in iteritems(indexers)]
+        indexers_list = self._validate_indexers(indexers)
 
         variables = OrderedDict()
         for name, var in iteritems(self._variables):
-            var_indexers = dict((k, v) for k, v in indexers if k in var.dims)
+            var_indexers = {k: v for k, v in indexers_list if k in var.dims}
             new_var = var.isel(**var_indexers)
             if not (drop and name in var_indexers):
                 variables[name] = new_var
-        coord_names = set(self._coord_names) & set(variables)
+
+        coord_names = set(variables).intersection(self._coord_names)
+        selected = self._replace_vars_and_dims(variables,
+                                               coord_names=coord_names)
+
+        # Extract coordinates from indexers
+        coord_vars = selected._get_indexers_coordinates(indexers)
+        variables.update(coord_vars)
+        coord_names = (set(variables)
+                       .intersection(self._coord_names)
+                       .union(coord_vars))
         return self._replace_vars_and_dims(variables, coord_names=coord_names)
 
     def sel(self, method=None, tolerance=None, drop=False, **indexers):
@@ -1217,26 +1295,45 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject,
             by scalars, slices or arrays of tick labels. For dimensions with
             multi-index, the indexer may also be a dict-like object with keys
             matching index level names.
+            If DataArrays are passed as indexers, xarray-style indexing will be
+            carried out. See :ref:`indexing` for the details.
 
         Returns
         -------
         obj : Dataset
             A new Dataset with the same contents as this dataset, except each
-            variable and dimension is indexed by the appropriate indexers. In
-            general, each variable's data will be a view of the variable's data
-            in this dataset, unless numpy fancy indexing was triggered by using
+            variable and dimension is indexed by the appropriate indexers.
+            If indexer DataArrays have coordinates that do not conflict with
+            this object, then these coordinates will be attached.
+            In general, each array's data will be a view of the array's data
+            in this dataset, unless vectorized indexing was triggered by using
             an array indexer, in which case the data will be a copy.
+
 
         See Also
         --------
         Dataset.isel
-        Dataset.sel_points
-        Dataset.isel_points
         DataArray.sel
         """
+        from .dataarray import DataArray
+
+        v_indexers = {k: v.variable.data if isinstance(v, DataArray) else v
+                      for k, v in indexers.items()}
+
         pos_indexers, new_indexes = indexing.remap_label_indexers(
-            self, indexers, method=method, tolerance=tolerance
+            self, v_indexers, method=method, tolerance=tolerance
         )
+        # attach indexer's coordinate to pos_indexers
+        for k, v in indexers.items():
+            if isinstance(v, Variable):
+                pos_indexers[k] = Variable(v.dims, pos_indexers[k])
+            elif isinstance(v, DataArray):
+                # drop coordinates found in indexers since .sel() already
+                # ensures alignments
+                coords = OrderedDict((k, v) for k, v in v._coords.items()
+                                     if k not in indexers)
+                pos_indexers[k] = DataArray(pos_indexers[k],
+                                            coords=coords, dims=v.dims)
         result = self.isel(drop=drop, **pos_indexers)
         return result._replace_indexes(new_indexes)
 
@@ -1278,6 +1375,8 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject,
         Dataset.sel_points
         DataArray.isel_points
         """
+        warnings.warn('Dataset.isel_points is deprecated: use Dataset.isel()'
+                      'instead.', DeprecationWarning, stacklevel=2)
 
         indexer_dims = set(indexers)
 
@@ -1424,6 +1523,9 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject,
         Dataset.isel_points
         DataArray.sel_points
         """
+        warnings.warn('Dataset.sel_points is deprecated: use Dataset.sel()'
+                      'instead.', DeprecationWarning, stacklevel=2)
+
         pos_indexers, _ = indexing.remap_label_indexers(
             self, indexers, method=method, tolerance=tolerance
         )