+++ xarray/core/variable.py
@@ -19,7 +19,9 @@ from . import ops
 from . import utils
 from .pycompat import (basestring, OrderedDict, zip, integer_types,
                        dask_array_type)
-from .indexing import (PandasIndexAdapter, orthogonally_indexable)
+from .indexing import (PandasIndexAdapter, xarray_indexable, BasicIndexer,
+                       OuterIndexer, VectorizedIndexer)
+from .utils import OrderedSet
 
 import xarray as xr  # only for Dataset and DataArray
 
@@ -27,6 +29,16 @@ try:
     import dask.array as da
 except ImportError:
     pass
+
+
+BASIC_INDEXING_TYPES = integer_types + (slice,)
+
+
+class MissingDimensionsError(ValueError):
+    """Error class used when we can't safely guess a dimension name.
+    """
+    # inherits from ValueError for backward compatibility
+    # TODO: move this to an xarray.exceptions module?
 
 
 def as_variable(obj, name=None):
@@ -87,7 +99,7 @@ def as_variable(obj, name=None):
     elif name is not None:
         data = as_compatible_data(obj)
         if data.ndim != 1:
-            raise ValueError(
+            raise MissingDimensionsError(
                 'cannot set variable %r with %r-dimensional data '
                 'without explicit dimension names. Pass a tuple of '
                 '(dims, data) instead.' % (name, data.ndim))
@@ -99,11 +111,11 @@ def as_variable(obj, name=None):
     if name is not None and name in obj.dims:
         # convert the Variable into an Index
         if obj.ndim != 1:
-            raise ValueError(
+            raise MissingDimensionsError(
                 '%r has more than 1-dimension and the same name as one of its '
                 'dimensions %r. xarray disallows such variables because they '
-                'conflict with the coordinates used to label dimensions.'
-                % (name, obj.dims))
+                'conflict with the coordinates used to label '
+                'dimensions.' % (name, obj.dims))
         obj = obj.to_index_variable()
 
     return obj
@@ -305,7 +317,7 @@ class Variable(common.AbstractArray, utils.NdimSizeLenMixin):
 
     @property
     def _indexable_data(self):
-        return orthogonally_indexable(self._data)
+        return xarray_indexable(self._data)
 
     def load(self, **kwargs):
         """Manually trigger loading of this variable's data from disk or a
@@ -404,34 +416,188 @@ class Variable(common.AbstractArray, utils.NdimSizeLenMixin):
         else:
             return key
 
+    def _broadcast_indexes(self, key):
+        """Prepare an indexing key for an indexing operation.
+
+        Parameters
+        -----------
+        key: int, slice, array, dict or tuple of integer, slices and arrays
+            Any valid input for indexing.
+
+        Returns
+        -------
+        dims: tuple
+            Dimension of the resultant variable.
+        indexers: IndexingTuple subclass
+            Tuple of integer, array-like, or slices to use when indexing
+            self._data. The type of this argument indicates the type of
+            indexing to perform, either basic, outer or vectorized.
+        new_order : Optional[Sequence[int]]
+            Optional reordering to do on the result of indexing. If not None,
+            the first len(new_order) indexing should be moved to these
+            positions.
+        """
+        key = self._item_key_to_tuple(key)  # key is a tuple
+        # key is a tuple of full size
+        key = indexing.expanded_indexer(key, self.ndim)
+        # Convert a scalar Variable as an integer
+        key = tuple([(k.data.item() if isinstance(k, Variable) and k.ndim == 0
+                      else k) for k in key])
+        if all(isinstance(k, BASIC_INDEXING_TYPES) for k in key):
+            return self._broadcast_indexes_basic(key)
+
+        self._validate_indexers(key)
+        # Detect it can be mapped as an outer indexer
+        # If all key is unlabeled, or
+        # key can be mapped as an OuterIndexer.
+        if all(not isinstance(k, Variable) for k in key):
+            return self._broadcast_indexes_outer(key)
+
+        # If all key is 1-dimensional and there are no duplicate labels,
+        # key can be mapped as an OuterIndexer.
+        dims = []
+        for k, d in zip(key, self.dims):
+            if isinstance(k, Variable):
+                if len(k.dims) > 1:
+                    return self._broadcast_indexes_vectorized(key)
+                dims.append(k.dims[0])
+            if not isinstance(k, integer_types):
+                dims.append(d)
+
+        if len(set(dims)) == len(dims):
+            return self._broadcast_indexes_outer(key)
+
+        return self._broadcast_indexes_vectorized(key)
+
+    def _broadcast_indexes_basic(self, key):
+        dims = tuple(dim for k, dim in zip(key, self.dims)
+                     if not isinstance(k, integer_types))
+        return dims, BasicIndexer(key), None
+
+    def _validate_indexers(self, key):
+        """ Make sanity checks """
+        for dim, k in zip(self.dims, key):
+            if isinstance(k, BASIC_INDEXING_TYPES):
+                pass
+            else:
+                if not isinstance(k, Variable):
+                    k = np.asarray(k)
+                    if k.ndim > 1:
+                        raise IndexError(
+                            "Unlabeled multi-dimensional array cannot be "
+                            "used for indexing: {}".format(k))
+                if k.dtype.kind == 'b':
+                    if self.shape[self.get_axis_num(dim)] != len(k):
+                        raise IndexError(
+                            "Boolean array size {0:d} is used to index array "
+                            "with shape {1:s}.".format(len(k),
+                                                       str(self.shape)))
+                    if k.ndim > 1:
+                        raise IndexError("{}-dimensional boolean indexing is "
+                                         "not supported. ".format(k.ndim))
+                    if getattr(k, 'dims', (dim, )) != (dim, ):
+                        raise IndexError(
+                            "Boolean indexer should be unlabeled or on the "
+                            "same dimension to the indexed array. Indexer is "
+                            "on {0:s} but the target dimension is "
+                            "{1:s}.".format(str(k.dims), dim))
+
+    def _broadcast_indexes_outer(self, key):
+        dims = tuple(k.dims[0] if isinstance(k, Variable) else dim
+                     for k, dim in zip(key, self.dims)
+                     if not isinstance(k, integer_types))
+        indexer = []
+        for k in key:
+            if isinstance(k, Variable):
+                k = k.data
+
+            if isinstance(k, BASIC_INDEXING_TYPES):
+                indexer.append(k)
+            else:
+                k = np.asarray(k)
+                indexer.append(k if k.dtype.kind != 'b' else np.flatnonzero(k))
+        return dims, OuterIndexer(indexer), None
+
+    def _nonzero(self):
+        """ Equivalent numpy's nonzero but returns a tuple of Varibles. """
+        # TODO we should replace dask's native nonzero
+        # after https://github.com/dask/dask/issues/1076 is implemented.
+        nonzeros = np.nonzero(self.data)
+        return tuple(Variable((dim), nz) for nz, dim
+                     in zip(nonzeros, self.dims))
+
+    def _broadcast_indexes_vectorized(self, key):
+        variables = []
+        out_dims_set = OrderedSet()
+        for dim, value in zip(self.dims, key):
+            if isinstance(value, slice):
+                out_dims_set.add(dim)
+            else:
+                variable = (value if isinstance(value, Variable) else
+                            as_variable(value, name=dim))
+                if variable.dtype.kind == 'b':  # boolean indexing case
+                    (variable,) = variable._nonzero()
+
+                variables.append(variable)
+                out_dims_set.update(variable.dims)
+
+        variable_dims = set()
+        for variable in variables:
+            variable_dims.update(variable.dims)
+
+        slices = []
+        for i, (dim, value) in enumerate(zip(self.dims, key)):
+            if isinstance(value, slice):
+                if dim in variable_dims:
+                    # We only convert slice objects to variables if they share
+                    # a dimension with at least one other variable. Otherwise,
+                    # we can equivalently leave them as slices aknd transpose
+                    # the result. This is significantly faster/more efficient
+                    # for most array backends.
+                    values = np.arange(*value.indices(self.sizes[dim]))
+                    variables.insert(i - len(slices), Variable((dim,), values))
+                else:
+                    slices.append((i, value))
+
+        try:
+            variables = _broadcast_compat_variables(*variables)
+        except ValueError:
+            raise IndexError("Dimensions of indexers mismatch: {}".format(key))
+
+        out_key = [variable.data for variable in variables]
+        out_dims = tuple(out_dims_set)
+        slice_positions = set()
+        for i, value in slices:
+            out_key.insert(i, value)
+            new_position = out_dims.index(self.dims[i])
+            slice_positions.add(new_position)
+
+        if slice_positions:
+            new_order = [i for i in range(len(out_dims))
+                         if i not in slice_positions]
+        else:
+            new_order = None
+
+        return out_dims, VectorizedIndexer(out_key), new_order
+
     def __getitem__(self, key):
         """Return a new Array object whose contents are consistent with
         getting the provided key from the underlying data.
 
-        NB. __getitem__ and __setitem__ implement "orthogonal indexing" like
-        netCDF4-python, where the key can only include integers, slices
-        (including `Ellipsis`) and 1d arrays, each of which are applied
-        orthogonally along their respective dimensions.
-
-        The difference does not matter in most cases unless you are using
-        numpy's "fancy indexing," which can otherwise result in data arrays
-        whose shapes is inconsistent (or just uninterpretable with) with the
-        variable's dimensions.
+        NB. __getitem__ and __setitem__ implement xarray-style indexing,
+        where if keys are unlabeled arrays, we index the array orthogonally
+        with them. If keys are labeled array (such as Variables), they are
+        broadcasted with our usual scheme and then the array is indexed with
+        the broadcasted key, like numpy's fancy indexing.
 
         If you really want to do indexing like `x[x > 0]`, manipulate the numpy
         array `x.values` directly.
         """
-        key = self._item_key_to_tuple(key)
-        key = indexing.expanded_indexer(key, self.ndim)
-        dims = tuple(dim for k, dim in zip(key, self.dims)
-                     if not isinstance(k, integer_types))
-        values = self._indexable_data[key]
-        # orthogonal indexing should ensure the dimensionality is consistent
-        if hasattr(values, 'ndim'):
-            assert values.ndim == len(dims), (values.ndim, len(dims))
-        else:
-            assert len(dims) == 0, len(dims)
-        return type(self)(dims, values, self._attrs, self._encoding,
+        dims, index_tuple, new_order = self._broadcast_indexes(key)
+        data = self._indexable_data[index_tuple]
+        if new_order:
+            data = np.moveaxis(data, range(len(new_order)), new_order)
+        return type(self)(dims, data, self._attrs, self._encoding,
                           fastpath=True)
 
     def __setitem__(self, key, value):
@@ -440,15 +606,24 @@ class Variable(common.AbstractArray, utils.NdimSizeLenMixin):
 
         See __getitem__ for more details.
         """
-        key = self._item_key_to_tuple(key)
-        if isinstance(self._data, dask_array_type):
-            raise TypeError("this variable's data is stored in a dask array, "
-                            'which does not support item assignment. To '
-                            'assign to this variable, you must first load it '
-                            'into memory explicitly using the .load() '
-                            'method or accessing its .values attribute.')
-        data = orthogonally_indexable(self._data)
-        data[key] = value
+        dims, index_tuple, new_order = self._broadcast_indexes(key)
+
+        if isinstance(value, Variable):
+            value = value.set_dims(dims).data
+
+        if new_order:
+            value = duck_array_ops.asarray(value)
+            if value.ndim > len(dims):
+                raise ValueError(
+                    'shape mismatch: value array of shape %s could not be'
+                    'broadcast to indexing result with %s dimensions'
+                    % (value.shape, len(dims)))
+
+            value = value[(len(dims) - value.ndim) * (np.newaxis,) +
+                          (Ellipsis,)]
+            value = np.moveaxis(value, new_order, range(len(new_order)))
+
+        self._indexable_data[index_tuple] = value
 
     @property
     def attrs(self):
@@ -789,8 +964,8 @@ class Variable(common.AbstractArray, utils.NdimSizeLenMixin):
 
         missing_dims = set(self.dims) - set(dims)
         if missing_dims:
-            raise ValueError('new dimensions must be a superset of existing '
-                             'dimensions')
+            raise ValueError('new dimensions %r must be a superset of '
+                             'existing dimensions %r' % (dims, self.dims))
 
         self_dims = set(self.dims)
         expanded_dims = tuple(
@@ -1251,12 +1426,13 @@ class IndexVariable(Variable):
         return self.copy(deep=False)
 
     def __getitem__(self, key):
-        key = self._item_key_to_tuple(key)
-        values = self._indexable_data[key]
-        if not hasattr(values, 'ndim') or values.ndim == 0:
-            return Variable((), values, self._attrs, self._encoding)
+        dims, index_tuple, new_order = self._broadcast_indexes(key)
+        values = self._indexable_data[index_tuple]
+        if getattr(values, 'ndim', 0) != 1:
+            # returns Variable rather than IndexVariable if multi-dimensional
+            return Variable(dims, values, self._attrs, self._encoding)
         else:
-            return type(self)(self.dims, values, self._attrs,
+            return type(self)(dims, values, self._attrs,
                               self._encoding, fastpath=True)
 
     def __setitem__(self, key, value):